+++ boilerplate/web_modules/Footer/index.js
@@ -1,4 +1,5 @@
 import React, { Component } from "react"
+import { Link } from "react-router"
 
 import styles from "./index.css"
 
@@ -16,6 +17,22 @@ export default class Footer extends Component {
             {  "<Statinamic />" }
           </span>
         </a>
+
+        { " | " }
+        { "Pages: " }
+        <Link
+          className={ styles.link }
+          to="/404.html"
+        >
+          { "404" }
+        </Link>
+        { ", " }
+        <Link
+          className={ styles.link }
+          to="/loading/"
+        >
+          { "Loading" }
+        </Link>
       </footer>
     )
   }