+++ __tests__/cli.js
@@ -1,35 +1,45 @@
 import test from "ava"
 import { join } from "path"
 
-import { denodeify as asyncify } from "promise"
 import { exec } from "child_process"
-
-const asyncExec = asyncify(exec)
 
 const target = join(__dirname, "..", "test-boilerplate")
 const execOpts = { cwd: target }
 
-const statinamic = "./node_modules/.bin/statinamic"
+const statinamic = "node ./node_modules/.bin/statinamic"
 
-test("should throw if a CLI flag is NOT recognized", async (t) => {
-  return (
-    asyncExec(`${ statinamic } start --lol`, execOpts)
-    .then(
-      () => t.fail(),
-      (err) => {
-        t.ok(err.message.indexOf("Unknown option `--lol'") > -1)
-      }
-    )
-  )
-})
+// cannot be done for now
+// see the mess in configurator which use process.argv directly
+// and since build.js is spawned, we got too many args and cannot use
+// yargs.strict()
+// https://github.com/MoOx/statinamic/issues/363
+// test.cb("should throw if a CLI flag is NOT recognized", (t) => {
+//   const child = exec(
+//     `${ statinamic } start --lol --open=false`, execOpts,
+//     (err) => {
+//       if (err) {
+//         clearTimeout(timeout)
+//         t.ok(err.message.indexOf("Unknown argument") > -1)
+//         t.end()
+//       }
+//     }
+//   )
+//
+//   const timeout = setTimeout(() => {
+//     child.kill()
+//     t.fail()
+//     t.end()
+//   }, 500)
+// })
 
 test.cb("should NOT throw if a CLI flag is recognized", (t) => {
   const child = exec(
-    `${ statinamic } start --devPort=4000`, execOpts,
+    `${ statinamic } start --devPort=4000 --open=false`, execOpts,
 
     // should die quickly...
     (err) => {
       if (err) {
+        console.log(err)
         clearTimeout(timeout)
         t.fail()
         t.end()
@@ -40,8 +50,39 @@ test.cb("should NOT throw if a CLI flag is recognized", (t) => {
   // ...or be ok quickly
   // (so we assume it's ok and kill the process, we don't need the actual build)
   const timeout = setTimeout(() => {
+    child.kill()
     t.pass()
     t.end()
-    child.kill()
   }, 500)
+})
+
+test.cb("should NOT throw if port is used", (t) => {
+  const app = require("express")()
+
+  const server = app.listen(8081, (err) => {
+    if (err) {
+      t.fail()
+      t.end()
+    }
+
+    const child = exec(
+      `${ statinamic } start --devPort=8081`, execOpts,
+
+      (err) => {
+        if (err) {
+          console.log(err)
+          clearTimeout(timeout)
+          t.fail()
+          t.end()
+        }
+      }
+    )
+
+    const timeout = setTimeout(() => {
+      child.kill()
+      server.close()
+      t.pass()
+      t.end()
+    }, 2000)
+  })
 })