+++ src/_utils/html-metas/__tests__/index.js
@@ -5,14 +5,14 @@ import url from "url"
 import htmlMetas, { defaultMetas } from "../"
 
 test("html default metas", (t) => {
-  t.same(
+  t.deepEqual(
     htmlMetas(),
     defaultMetas
   )
 })
 
 test("html static metas", (t) => {
-  t.same(
+  t.deepEqual(
     htmlMetas({
       baseUrl: url.parse("http://domain.ext/"),
       css: [ "statinamic-client.css" ],
@@ -25,7 +25,7 @@ test("html static metas", (t) => {
 })
 
 test("html static metas with path", (t) => {
-  t.same(
+  t.deepEqual(
     htmlMetas({
       baseUrl: url.parse("http://domain.ext/basep/"),
       css: [ "statinamic-client.css" ],